Test building before running (#13)
It turns out go run . inside a docker container on a cloud provider is really slow, perhaps due to poor IO on the networked disks we're using? The build completes in 5 seconds on my machine, 20 seconds in the docker Linux VM on my machine, and over a minute on cloud providers (Azure and GKE were similar). We probably need to rethink our approach to running services here.

* Test building before running

* Update helloworld example to go-micro v2

* Bump go-micro to master